+++ tourcount/src/main/java/com/wmstein/tourcount/CountingActivity.java
@@ -13,9 +13,12 @@ import android.graphics.Typeface;
 import android.media.Ringtone;
 import android.media.RingtoneManager;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.PowerManager;
+import android.os.VibrationEffect;
+import android.os.Vibrator;
 import android.provider.MediaStore;
 import android.util.Log;
 import android.view.Menu;
@@ -53,6 +56,10 @@ import java.util.Objects;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.NavUtils;
 import androidx.core.content.ContextCompat;
+import androidx.work.Data;
+import androidx.work.OneTimeWorkRequest;
+import androidx.work.WorkManager;
+import androidx.work.WorkRequest;
 
 /****************************************************************************************
  * CountingActivity is the central activity of TourCount in portrait mode. 
@@ -65,7 +72,7 @@ import androidx.core.content.ContextCompat;
  <p>
  * Basic counting functions created by milo for BeeCount on 05/05/2014.
  * Adopted, modified and enhanced for TourCount by wmstein since 2016-04-18,
- * last modification on 2023-05-27
+ * last modification on 2023-06-10
  */
 public class CountingActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener, PermissionsDialogFragment.PermissionsGrantedCallback
 {
@@ -106,7 +113,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     private boolean fontPref;
     private boolean lhandPref; // true for lefthand mode of counting screen
     private boolean buttonSoundPref;
-    private String buttonAlertSound;
+    private boolean buttonVibPref;
+    private String buttonSound;
+    private String buttonSoundMinus;
     private boolean metaPref;      // option for reverse geocoding
     private String emailString = ""; // mail address for OSM query
 
@@ -135,7 +144,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
         prefs.registerOnSharedPreferenceChangeListener(this);
         getPrefs();
 
-        if (lhandPref) // if left-handed counting page
+        // if left-handed counting page
+        if (lhandPref)
         {
             setContentView(R.layout.activity_counting_lh);
             LinearLayout counting_screen = findViewById(R.id.countingScreenLH);
@@ -198,8 +208,10 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
         sortPref = prefs.getString("pref_sort_sp", "none");    // sorted species list on counting page
         fontPref = prefs.getBoolean("pref_note_font", false);  // larger font for remarks
         lhandPref = prefs.getBoolean("pref_left_hand", false); // left-handed counting page
-        buttonSoundPref = prefs.getBoolean("pref_button_sound", false);
-        buttonAlertSound = prefs.getString("alert_button_sound", null);
+        buttonSoundPref = prefs.getBoolean("pref_button_sound", false); // make button sound
+        buttonVibPref = prefs.getBoolean("pref_button_vib", false); // make vibration
+        buttonSound = prefs.getString("button_sound", null); // use standard button sound
+        buttonSoundMinus = prefs.getString("button_sound_minus", null); //use deeper button sound
         metaPref = prefs.getBoolean("pref_metadata", false);   // use Reverse Geocoding
         emailString = prefs.getString("email_String", "");     // for reliable query of Nominatim service
         itemPosition = prefs.getInt("item_Position", 0);       // spinner pos.
@@ -402,13 +414,21 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
         // get reverse geocoding
         if (locationService.canGetLocation() && metaPref && (latitude != 0 || longitude != 0))
         {
-            // Trial with IntendService
             String urlString = "https://nominatim.openstreetmap.org/reverse?email=" + emailString
                 + "&format=xml&lat=" + latitude + "&lon=" + longitude + "&zoom=18&addressdetails=1";
 
-            Intent rintent = new Intent(this, RetrieveAddrService.class);
-            rintent.putExtra("urlString", urlString);
-            startService(rintent);
+            // Trial with WorkManager
+            WorkRequest retrieveAddrWorkRequest =
+                new OneTimeWorkRequest.Builder(RetrieveAddrWorker.class)
+                    .setInputData(new Data.Builder()
+                            .putString("URL_STRING", urlString)
+                            .build()
+                                 )
+                    .build();
+
+            WorkManager
+                .getInstance(this)
+                .enqueue(retrieveAddrWorkRequest);
         }
     }
 
@@ -538,6 +558,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpf1i(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 1; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -590,6 +612,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHf1i(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 1;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -640,7 +664,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownf1i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -677,7 +703,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHf1i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -714,6 +742,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpf2i(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 2; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -766,6 +796,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHf2i(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 2;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -816,7 +848,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownf2i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -852,7 +886,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHf2i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -886,6 +922,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpf3i(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 3; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -938,6 +976,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHf3i(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 3;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -988,7 +1028,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownf3i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1024,7 +1066,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHf3i(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1058,6 +1102,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUppi(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 4; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -1110,6 +1156,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHpi(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 4;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -1160,7 +1208,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownpi(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1196,7 +1246,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHpi(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1230,6 +1282,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpli(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 5; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -1282,6 +1336,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHli(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 5;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -1332,7 +1388,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownli(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1368,7 +1426,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHli(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1402,6 +1462,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpei(View view)
     {
         buttonSound();
+        buttonVib();
+
         // iAtt used by EditIndividualActivity to decide where to store bulk count value
         int iAtt = 6; // 1 f1i, 2 f2i, 3 f3i, 4 pi, 5 li, 6 ei
 
@@ -1454,6 +1516,8 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     public void countUpLHei(View view)
     {
         buttonSound();
+        buttonVib();
+
         int iAtt = 6;
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
@@ -1504,7 +1568,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownei(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidget widget = getCountFromId(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1540,7 +1606,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
     // deletes last count
     public void countDownLHei(View view)
     {
-        buttonSound();
+        buttonSoundMinus();
+        buttonVibLong();
+
         int count_id = Integer.parseInt(view.getTag().toString());
         CountingWidgetLH widget = getCountFromIdLH(count_id);
         spec_name = Objects.requireNonNull(widget).count.name; // set spec_name for toast in deleteIndividual
@@ -1655,9 +1723,9 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
             try
             {
                 Uri notification;
-                if (isNotBlank(buttonAlertSound) && buttonAlertSound != null)
+                if (isNotBlank(buttonSound) && buttonSound != null)
                 {
-                    notification = Uri.parse(buttonAlertSound);
+                    notification = Uri.parse(buttonSound);
                 }
                 else
                 {
@@ -1669,6 +1737,71 @@ public class CountingActivity extends AppCompatActivity implements SharedPrefere
             {
                 if (MyDebug.LOG)
                     Log.e(TAG, "could not play botton sound.", e);
+            }
+        }
+    }
+
+    private void buttonSoundMinus()
+    {
+        if (buttonSoundPref)
+        {
+            try
+            {
+                Uri notification;
+                if (isNotBlank(buttonSoundMinus) && buttonSoundMinus != null)
+                {
+                    notification = Uri.parse(buttonSoundMinus);
+                }
+                else
+                {
+                    notification = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+                }
+                Ringtone r = RingtoneManager.getRingtone(getApplicationContext(), notification);
+                r.play();
+            } catch (Exception e)
+            {
+                if (MyDebug.LOG)
+                    Log.e(TAG, "could not play botton sound.", e);
+            }
+        }
+    }
+
+    private void buttonVib()
+    {
+        if (buttonVibPref)
+        {
+            try
+            {
+                Vibrator vibrator = (Vibrator) getSystemService(VIBRATOR_SERVICE);
+                if (Build.VERSION.SDK_INT >= 26) {
+                    vibrator.vibrate(VibrationEffect.createOneShot(100, VibrationEffect.DEFAULT_AMPLITUDE));
+                } else {
+                    vibrator.vibrate(100);
+                }
+            } catch (Exception e)
+            {
+                if (MyDebug.LOG)
+                    Log.e(TAG, "could not vibrate.", e);
+            }
+        }
+    }
+
+    private void buttonVibLong()
+    {
+        if (buttonVibPref)
+        {
+            try
+            {
+                Vibrator vibrator = (Vibrator) getSystemService(VIBRATOR_SERVICE);
+                if (Build.VERSION.SDK_INT >= 26) {
+                    vibrator.vibrate(VibrationEffect.createOneShot(450, VibrationEffect.DEFAULT_AMPLITUDE));
+                } else {
+                    vibrator.vibrate(450);
+                }
+            } catch (Exception e)
+            {
+                if (MyDebug.LOG)
+                    Log.e(TAG, "could not vibrate.", e);
             }
         }
     }